Reorganise Schema classes in separate files
Add type annotations to public methods and values
